wip asdf refactor
