+++ test/Test/Asdf/DecodeSpec.hs
@@ -12,8 +12,8 @@ import Skeletest.Predicate qualified as P
 import Telescope.Asdf.Class
 import Telescope.Asdf.Core
 import Telescope.Asdf.Decoding
+import Telescope.Asdf.Encoding.File
 import Telescope.Asdf.Error
-import Telescope.Asdf.File
 import Telescope.Asdf.Node
 import Telescope.Asdf.Parser
 import Test.Asdf.FileSpec (ExampleFileFix (..))
